clean react arthitecture is cretaeted
