+++ front-end/src/App.js
@@ -1,6 +1,6 @@
 import "./App.css";
-import Footer from "./components/Footer";
-import Header from "./components/Header";
+import Footer from "./components/footer/footer";
+import Header from "./components/header/header";
 import MainRoutes from "./routes/MainRoutes";
 
 function App() {
