--- front-end/src/components/SignupModal.jsx
@@ -1,134 +1,0 @@
-import React from "react";
-import Button from "@material-ui/core/Button";
-import CssBaseline from "@material-ui/core/CssBaseline";
-import TextField from "@material-ui/core/TextField";
-import Link from "@material-ui/core/Link";
-import Grid from "@material-ui/core/Grid";
-import Typography from "@material-ui/core/Typography";
-import { makeStyles } from "@material-ui/core/styles";
-import Container from "@material-ui/core/Container";
-import SlackLoginButton from "../components/auth-components/SlackLoginButton"
-import { useNavigate } from "react-router-dom";
-
-const useStyles = makeStyles(theme => ({
-    "@global": {
-        body: {
-            backgroundColor: theme.palette.common.white
-        }
-    },
-    paper: {
-        marginTop: theme.spacing(8),
-        display: "flex",
-        flexDirection: "column",
-        alignItems: "center"
-    },
-    avatar: {
-        margin: theme.spacing(1),
-        backgroundColor: theme.palette.secondary.main
-    },
-    form: {
-        width: "100%", // Fix IE 11 issue.
-        marginTop: theme.spacing(3)
-    },
-    submit: {
-        margin: theme.spacing(3, 0, 2)
-    },
-    customStyle: {
-        backgroundColor: 'rgb(237, 217, 217)',
-        padding: theme.spacing(2),
-        marginRight: "30px",
-        marginTop: "100px",
-    },
-}));
-
-
-export default function SignupModal() {
-    const classes = useStyles();
-    const navigate = useNavigate();
-
-    const handleSignup = () => {
-        navigate("/main");
-    }
-    
-    return (
-        <Container component="main" maxWidth="xs" className={classes.customStyle}>
-            <CssBaseline />
-            <div className={classes.paper}>
-                <Typography component="h1" variant="h4">
-                    Welcome to Our Platform!
-                </Typography>
-                <Typography component="h4" variant="h6">
-                    Explore our courses and upcoming classes.
-                </Typography>
-                <form className={classes.form} noValidate>
-                    <Grid container spacing={2}>
-                        <Grid item xs={12} sm={6}>
-                            <TextField
-                                autoComplete="fname"
-                                name="firstName"
-                                variant="outlined"
-                                required
-                                fullWidth
-                                id="firstName"
-                                label="First Name"
-                                autoFocus
-                            />
-                        </Grid>
-                        <Grid item xs={12} sm={6}>
-                            <TextField
-                                variant="outlined"
-                                required
-                                fullWidth
-                                id="lastName"
-                                label="Last Name"
-                                name="lastName"
-                                autoComplete="lname"
-                            />
-                        </Grid>
-                        <Grid item xs={12}>
-                            <TextField
-                                variant="outlined"
-                                required
-                                fullWidth
-                                id="email"
-                                label="Email Address"
-                                name="email"
-                                autoComplete="email"
-                            />
-                        </Grid>
-                        <Grid item xs={12}>
-                            <TextField
-                                variant="outlined"
-                                required
-                                fullWidth
-                                name="password"
-                                label="Password"
-                                type="password"
-                                id="password"
-                                autoComplete="current-password"
-                            />
-                        </Grid>
-                    </Grid>
-                    <Button
-                        type="submit"
-                        fullWidth
-                        variant="contained"
-                        color="primary"
-                        className={classes.submit}
-                        onClick={handleSignup}
-                    >
-                        Sign Up
-                    </Button>
-                    <SlackLoginButton/>
-                    <Grid container justify="flex-end">
-                        <Grid item>
-                            <Link href="/" variant="body2">
-                                Already have an account? Sign in
-                            </Link>
-                        </Grid>
-                    </Grid>
-                </form>
-            </div>
-        </Container>
-    );
-}
