+++ front-end/src/components/auth-components/LoginForm.jsx
@@ -7,10 +7,9 @@ import FormProvider from "../hooks-form/form-provider";
 import RHFTextField from "../hooks-form/RHFTextField";
 import { Button, Divider } from "@mui/material";
 import SlackLoginButton from "./SlackLoginButton";
-const LoginForm = () => {
+const LoginForm = ({onSubmit}) => {
 
-  //const url = "https://localhost:443/auth/redirect"
-  const oAuthUrl = "https://localhost:443/auth"
+
 
   const schema = Yup.object().shape({
     username: Yup.string().max(255, "Max 255").required(),
@@ -39,21 +38,10 @@ const LoginForm = () => {
     formState: { isSubmitting, isValid },
   } = methods;
 
-  const onSubmit = (data) => {
-    console.log("submit", data);
-  }
-  const handlerFailure = (error) => {
-    console.log({ error })
-  }
 
-  const handlerSuccess = async (code) => {
-    // The component will return a slack OAuth verifier code, 
-    // you should send that code to your API and exchange your temporary OAuth verifier code for an access token.
-    const request = await fetch(oAuthUrl, { method: 'POST', headers: { 'Content-Type': 'application/json' }, body: JSON.stringify({ code }) })
-    // If successfull you can redirect the user to the app.
-    const json = await request.json()
-    console.log(json)
-  }
+
+
+
 
   return (
     <FormProvider methods={methods} onSubmit={handleSubmit(onSubmit)}>
@@ -63,8 +51,6 @@ const LoginForm = () => {
         <RHFTextField name='password' label='Password' />
         <Button type='submit' variant="contained">Login</Button>
         <Divider />
-        <SlackLoginButton />
-        <Button href={`/signup`} >Sign Up</Button>
       </Stack>
     </FormProvider>
