+++ front-end/src/components/footer/footer.jsx
@@ -1,4 +1,4 @@
-import React from "react"
+import React from "react";
 
 const Footer = () => {
     return (
@@ -27,4 +27,4 @@ const Footer = () => {
     );
 }
 
-export default Footer; 
+export default Footer;
