--- front-end/src/pages/signup/signup.jsx
@@ -1,17 +1,0 @@
-import React from 'react';
-import { Container } from '@mui/material';
-import SignUp from '../../components/auth-components/SignUpForm';
-
-
-const SignUpPage = () => {
-
-  
-  return (
-<Container maxWidth="sm">
-      <h2>Sign Up Page</h2>
-      <SignUp />
-   </Container>
-  );
-}
-
-export default SignUpPage;
