+++ front-end/src/routes/MainRoutes.js
@@ -1,13 +1,16 @@
 import React from "react";
 import { Routes, Route } from "react-router-dom";
-import HomePage from "../components/HomaPage.jsx"
-import Main from "../components/Main.jsx";
+import Main from "../pages/main";
+import SignUp from "../pages/signup";
+import Login from "../pages/login";
+
 
 const MainRoutes = () => {
   return (
     <Routes>
-      <Route path="/" element={<HomePage />} />
+      <Route path="/" element={<SignUp />} />
       <Route path="/main" element={<Main />}/>
+      <Route path='/login' element={<Login />}/>
     </Routes>
   );
 };
